CategoryRepository moved to parent package
